SocialFeed Creation
A very rough start for the social feed and post making portion of the Code Together project.
This was made by ai (Gemini Pro model) since it's been forever since the last time I have touched html, but it's a good start. I've added on and made changes to the base I was given, along with integrating Parker's cool Matrix javascript stuff.
It is all currently in one html file, will eventually make the java script and css file. Would be good to decide on a semi-universal style sheet for new pages.
-Also created a subfolder to keep the post related stuff in.