Profile Page, Home Page + extra
Added a basic layout and functionalities for the Profile Page, Home Page as well as an experimental 3d environment from imported js code, though we may not use that.
This update includes the main html and php files along with js scripts.
The profile page allows for making posts and saving them locally for now, it also as way to talk with an ai via OpenAi's api, may not work on other computers yet lol.
Need to work on the homepage next, especially with the connection between it and the server.