AI Update - More chars + Modular
Added Agent Smith and the Oracle as ai characters. Made a client and server side array of prompts, for the potential of the user making their own ai assistant.
Made the ai.php more modular so it should work on other pages without much work, hopefully it will be fine...
Also added a few images for ai characters.