updated
changed the winner box to be a modal so when timer is up winner pops up. itll need some love from back end to display the correct winning group. Im pretty sure i found the error for the color problems and border issue for teama and teamb so I changed that up to.