Made Posts modular + delete posts
Posts are now in their own php in the includes folder, just run a for loop with the includes inside of it from any page. Got annoying getting the home and profile posts consistent.
Added drop down that has save post, report post, and delete post. Only delete works for now.
Added the ability to 'soft' delete your own posts. It just uses the deletePost() method that marks the post as deleted, but is still technically in the database but won't show up in calls from DAO that have 'is_deleted = FALSE' in it's query.