Added LoadingScreen with fading transitions
A simple loading screen that will fade in and out when switching to a different page. Currently you have to specify the fade in, all of the nav bar links work with it right now.
Tell me if this is annoying or if it's not needed, kinda like it better than it just moving instantly from page to page as it feels more modern.