DEATH TO LIKE POST REDIRECTS!!!
Fixed the like button from reloading the page. Was annoying because it triggered the loading screen and brought you to the top of the screen with the whole page refreshed.

To solve this problem I simply nuked the LikePostController with some AJAX from post.js. Honestly works pretty good, I don't know if there are any downsides from this approach but I like it A LOT better than how it was at least.